Button Functions
Added basic functionality to the buttons so that we can at least move between the different sections of the app. For the time being, this is just acting as if the user logged in as a client, since the driver main screen hasn't been made yet.

So right now, you can log in as foo@example.com with the password hello (it's hardcoded until we can work out how we program it to check the database for account credentials), go into and out of settings, and log out (i.e. go back to the login screen).

It's suppose to also allow you to view the Request activity page, but for some reason it crashes when doing so. I still can't figure out why.