Finishing Map-related Changes
When a request is active, the Customer will have a map on their main screen that will allow them to track their driver. It refreshes every 3 minutes.

For the Driver, two extra buttons will appear on their screen. These buttons will open Google Maps to give the Driver directions to either the pick-up or drop-off location, depending on which button they clicked on.

For both pages, the map/buttons will be hidden if a request is not active.