Checking Active Request and Map Improvements
Added in a call to the checkForActiveSRById function to check if the currently logged in user (Customer or Driver) is part of an active request. If so, they should see the appropriate changes to their screen.

Also got the map set up so that it at least tries to get the user's current location. If it can't (due to permission issues or something else), then it will set it to roughly the classroom's location by default.